Feature update delete

+++ server/routes/shelf.router.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const pool = require('../modules/pool');
 const router = express.Router();
+const { rejectUnauthenticated } = require('../modules/authentication-middleware');
 
 /**
  * Get all of the items on the shelf
@@ -17,7 +18,6 @@ router.get('/', (req, res) => {
             res.sendStatus(500);
         })
 });
-
 
 /**
  * Add an item for the logged in user to the shelf
@@ -43,17 +43,23 @@ router.post('/', (req, res) => {
 /**
  * Delete an item if it's something the logged in user added
  */
-router.delete('/:id', (req, res) => {
-    console.log(req.params)
+router.delete('/:id/:user_id', rejectUnauthenticated, (req, res) => {
+    console.log(req.user.id)
     const id = req.params.id
+    const user_id = req.params.user_id
+    const loggedin_user = req.user.id
     console.log('in delete route', id)
-    const queryText = 'DELETE FROM "item" WHERE "id" = $1'
-    pool.query(queryText, [id])
-    .then(() => {res.sendStatus(200)})
-    .catch((err) => {
+    if (loggedin_user == user_id) {
+        const queryText = 'DELETE FROM "item" WHERE "id" = $1'
+        pool.query(queryText, [id])
+        .then(() => {res.sendStatus(200)})
+        .catch((err) => {
         console.log(err)
         res.sendStatus(500)
     })
+    } else {
+        res.sendStatus(403)
+    }
 });
